Working tests with a compatible API
