+++ crypto-shared/Cargo.toml
@@ -4,13 +4,15 @@ edition = "2021"
 
 [dependencies]
 k256 = { version = "0.13.1", features = ["sha256", "ecdsa", "serde", "arithmetic", "expose-field"] }
-hkdf = "0.12.4"
 anyhow = "1"
 serde = "1"
 borsh = "1.3.0"
 near-account-id = "1"
-near-crypto = "0.21.2"
 serde_json = "1"
 near-sdk = "=5.0.0-alpha.1"
+
+# TODO ENSURE THIS ISN'T BUILT INTO THE NODE
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+getrandom = { version = "0.2.12", features = ["custom"] }
 
 [dev-dependencies]
